Refactor site structure and refactor LoginTest
