+++ src/main/java/net/ficbook/irinanovik/driver/DriverSingleton.java
@@ -1,4 +1,4 @@
-package net.ficbook.irinanovik;
+package net.ficbook.irinanovik.driver;
 
 import org.openqa.selenium.WebDriver;
 import org.openqa.selenium.chrome.ChromeDriver;
