+++ src/main/java/net/ficbook/irinanovik/pages/CookiePage.java
@@ -1,5 +1,7 @@
-package net.ficbook.irinanovik;
+package net.ficbook.irinanovik.pages;
 
+import net.ficbook.irinanovik.driver.DriverSingleton;
+import net.ficbook.irinanovik.utils.Waiters;
 import org.openqa.selenium.By;
 import org.openqa.selenium.WebDriver;
 import org.openqa.selenium.WebElement;
