+++ src/main/java/net/ficbook/irinanovik/pages/CookiePageXPath.java
@@ -1,4 +1,4 @@
-package net.ficbook.irinanovik;
+package net.ficbook.irinanovik.pages;
 
 public class CookiePageXPath {
     public static final String BUTTON_CLOSE_COOKIE_XPATH = "/html/body/div[1]/div/div[2]/div/div[1]/div[2]/form/button";
