+++ src/main/java/net/ficbook/irinanovik/pages/HomePageXPath.java
@@ -1,4 +1,4 @@
-package net.ficbook.irinanovik;
+package net.ficbook.irinanovik.pages;
 
 public class HomePageXPath {
     public static final String BUTTON_LOGIN_XPATH = "document.querySelector(\"body > div.js-modal-destination > div > div.modal-dialog > div > div.modal-body > div.modal-container > form > button\")";
