+++ src/main/java/net/ficbook/irinanovik/pages/LoginPage.java
@@ -1,4 +1,4 @@
-package net.ficbook.irinanovik;
+package net.ficbook.irinanovik.pages;
 
 public class LoginPage {
