+++ src/test/java/net/ficbook/irinanovik/LoginTest.java
@@ -1,15 +1,16 @@
 package net.ficbook.irinanovik;
 
+import net.ficbook.irinanovik.driver.DriverSingleton;
+import net.ficbook.irinanovik.pages.HomePage;
+import net.ficbook.irinanovik.pages.LoginPage;
 import org.junit.jupiter.api.AfterEach;
 import org.junit.jupiter.api.Assertions;
 import org.junit.jupiter.api.Test;
-import org.openqa.selenium.WebDriver;
-import org.openqa.selenium.chrome.ChromeDriver;
 
 public class LoginTest {
 
     @Test
-    public void testLoginWithNotExistTextOfEmailAndPassword() {
+    public void testLoginWithNotExistTextOfUsernameAndPassword() {
 
         HomePage homePage = new HomePage();
         homePage.open();
@@ -25,6 +26,24 @@ public class LoginTest {
 
         Assertions.assertEquals(expectedTextPassword,actualTextPassword);
     }
+    @Test
+    public void testLoginWithEmptyPassword() {
+
+        HomePage homePage = new HomePage();
+        homePage.open();
+        homePage.clickButtonLogin();
+
+        LoginPage loginPage = new LoginPage();
+        loginPage.inputUsername("fdgbzfd");
+        loginPage.inputPassword("");
+        loginPage.clickButtonEnter();
+
+        String actualTextPassword = loginPage.errorPasswordText().getText();
+        String expectedTextPassword = "Логин не существует";
+
+        Assertions.assertEquals(expectedTextPassword,actualTextPassword);
+    }
+}
 
     @AfterEach
     public void close() {
